Conflicts:
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/StringMessages.java
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/StringMessages.properties
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/client/StringMessages_de.properties
	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/raceboard/TouchSplitLayoutPanel.java

+++ java/com.sap.sse.landscape.aws/src/com/sap/sse/landscape/aws/orchestration/CopyAndCompareMongoDatabase.java
@@ -27,13 +27,15 @@ extends AbstractAwsProcedureImpl<ShardingKey> {
     private final Database targetDatabase;
     private final Iterable<Database> additionalDatabasesToDelete;
     private final boolean dropTargetFirst;
+    private final boolean dropSourceAfterSuccessfulCopy;
 
     public static interface Builder<BuilderT extends AbstractAwsProcedureImpl.Builder<BuilderT, CopyAndCompareMongoDatabase<ShardingKey>, ShardingKey>, ShardingKey>
     extends AbstractAwsProcedureImpl.Builder<BuilderT, CopyAndCompareMongoDatabase<ShardingKey>, ShardingKey> {
         BuilderT setSourceDatabase(Database sourceDatabase);
         BuilderT setTargetDatabase(Database targetDatabase);
         BuilderT setAdditionalDatabasesToDelete(Iterable<Database> additionalDatabasesToDelete);
-        BuilderT dropTargetFirst(boolean b);
+        BuilderT dropTargetFirst(boolean dropTargetFirst);
+        BuilderT dropSourceAfterSuccessfulCopy(boolean dropSourceAfterSuccessfulCopy);
     }
     
     public static class BuilderImpl<BuilderT extends Builder<BuilderT, ShardingKey>, ShardingKey>
@@ -43,6 +45,7 @@ extends AbstractAwsProcedureImpl<ShardingKey> {
         private Database targetDatabase;
         private Iterable<Database> additionalDatabasesToDelete;
         private boolean dropTargetFirst;
+        private boolean dropSourceAfterSuccessfulCopy;
         
         @Override
         public CopyAndCompareMongoDatabase<ShardingKey> build() throws Exception {
@@ -73,6 +76,12 @@ extends AbstractAwsProcedureImpl<ShardingKey> {
             return self();
         }
 
+        @Override
+        public BuilderT dropSourceAfterSuccessfulCopy(boolean dropSourceAfterSuccessfulCopy) {
+            this.dropSourceAfterSuccessfulCopy = dropSourceAfterSuccessfulCopy;
+            return self();
+        }
+
         protected Database getSourceDatabase() {
             return sourceDatabase;
         }
@@ -88,6 +97,10 @@ extends AbstractAwsProcedureImpl<ShardingKey> {
         protected boolean isDropTargetFirst() {
             return dropTargetFirst;
         }
+        
+        protected boolean isDropSourceAfterSuccessfulCopy() {
+            return dropSourceAfterSuccessfulCopy;
+        }
     }
     
     public static <BuilderT extends Builder<BuilderT, ShardingKey>, ShardingKey>
@@ -101,6 +114,7 @@ extends AbstractAwsProcedureImpl<ShardingKey> {
         this.targetDatabase = builder.getTargetDatabase();
         this.additionalDatabasesToDelete = builder.getAdditionalDatabasesToDelete() == null ? Collections.emptySet() : builder.getAdditionalDatabasesToDelete();
         this.dropTargetFirst = builder.isDropTargetFirst();
+        this.dropSourceAfterSuccessfulCopy = builder.isDropSourceAfterSuccessfulCopy();
     }
 
     @Override
@@ -121,10 +135,13 @@ extends AbstractAwsProcedureImpl<ShardingKey> {
         final Future<String> targetMd5 = executor.submit(()->targetDatabase.getMD5Hash());
         executor.shutdown();
         if (sourceMd5.get().equals(targetMd5.get())) {
-            logger.info("Databases "+sourceDatabase+" and "+targetDatabase+" have equal MD5 hash "+sourceMd5.get()+". Removing "+sourceDatabase+" and "+Util.joinStrings(", ", additionalDatabasesToDelete));
-            sourceDatabase.drop();
-            for (final Database additionalDatabaseToDelete : additionalDatabasesToDelete) {
-                additionalDatabaseToDelete.drop();
+            logger.info("Databases "+sourceDatabase+" and "+targetDatabase+" have equal MD5 hash "+sourceMd5.get()+".");
+            if (dropSourceAfterSuccessfulCopy) {
+                logger.info("Removing "+sourceDatabase+" and "+Util.joinStrings(", ", additionalDatabasesToDelete));
+                sourceDatabase.drop();
+                for (final Database additionalDatabaseToDelete : additionalDatabasesToDelete) {
+                    additionalDatabaseToDelete.drop();
+                }
             }
         } else {
             throw new IllegalStateException("Import failed; hashes are different. "+sourceDatabase+" has "+sourceMd5.get()+
